# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Auction System",
        "description": "Auction System API Documentation",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/v1/images/upload": {
            "post": {
                "tags": [
                    "Images",
                    "Image Management"
                ],
                "summary": "Upload Image",
                "description": "Upload image to local storage\n\nPOST /images/upload\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nForm Data:\n- file: Image file (JPEG, PNG, WEBP)\n- product_id: Optional product ID for organization\n\nReturns: Uploaded image information with local path",
                "operationId": "upload_image_api_v1_images_upload_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_image_api_v1_images_upload_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Upload Image Api V1 Images Upload Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/images/upload/multiple": {
            "post": {
                "tags": [
                    "Images",
                    "Image Management"
                ],
                "summary": "Upload Multiple Images",
                "description": "Upload multiple images to local storage\n\nPOST /images/upload/multiple\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nForm Data:\n- files: List of image files (max 5)\n- product_id: Optional product ID for organization\n\nReturns: List of uploaded images with their information",
                "operationId": "upload_multiple_images_api_v1_images_upload_multiple_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_multiple_images_api_v1_images_upload_multiple_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Upload Multiple Images Api V1 Images Upload Multiple Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/images/delete": {
            "delete": {
                "tags": [
                    "Images",
                    "Image Management"
                ],
                "summary": "Delete Uploaded Image",
                "description": "Delete image from local storage\n\nDELETE /images/delete?image_path=storage/images/products/sample.jpg\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nQuery Parameters:\n- image_path: Relative path to image to delete\n\nReturns: Success message",
                "operationId": "delete_uploaded_image_api_v1_images_delete_delete",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "image_path",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Image Path"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Delete Uploaded Image Api V1 Images Delete Delete"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/images/view/{image_path}": {
            "get": {
                "tags": [
                    "Images",
                    "Image Management"
                ],
                "summary": "View Image",
                "description": "Serve image files from local storage\n\nGET /images/view/storage/images/products/sample.jpg\n\nReturns: Image file",
                "operationId": "view_image_api_v1_images_view__image_path__get",
                "parameters": [
                    {
                        "name": "image_path",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Image Path"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/images/list": {
            "get": {
                "tags": [
                    "Images",
                    "Image Management"
                ],
                "summary": "List Images",
                "description": "List images in storage with formatted sizes\n\nGET /images/list?product_id=123\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nQuery Parameters:\n- product_id: Optional filter by product ID\n\nReturns: List of images with their information",
                "operationId": "list_images_api_v1_images_list_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Product Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response List Images Api V1 Images List Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/images/samples": {
            "post": {
                "tags": [
                    "Images",
                    "Image Management"
                ],
                "summary": "Create Sample Images",
                "description": "Create sample images for demo purposes (Admin only)\n\nPOST /images/samples\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\n\nReturns: Sample images creation result",
                "operationId": "create_sample_images_api_v1_images_samples_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Create Sample Images Api V1 Images Samples Post"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_upload_image_api_v1_images_upload_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    },
                    "product_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Product Id"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_image_api_v1_images_upload_post"
            },
            "Body_upload_multiple_images_api_v1_images_upload_multiple_post": {
                "properties": {
                    "files": {
                        "items": {
                            "type": "string",
                            "format": "binary"
                        },
                        "type": "array",
                        "title": "Files"
                    },
                    "product_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Product Id"
                    }
                },
                "type": "object",
                "required": [
                    "files"
                ],
                "title": "Body_upload_multiple_images_api_v1_images_upload_multiple_post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}